% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maille.R
\name{get_maille}
\alias{get_maille}
\title{Extract cell coordinate}
\source{
Meteo France -> meteo.data.gouv.fr et Bertuzzi, Patrick; Clastre,
Philippe, 2022, "Information sur les mailles SAFRAN", DOI 10.57745/1PDFNL,
Recherche Data Gouv, V2.
}
\usage{
get_maille(geom)
}
\arguments{
\item{geom}{-> Polygone ou point pour lequel on veut obtenir les
coordonnees de la maille Safran.}
}
\value{
La fonction renvoie un data frame contenant les coordonnees X et Y
en Lambert II étendu et correspondant au polygone ou point d'entre.
}
\description{
Cette fonction permet d'extraire les coordonnees de la ou des
mailles correspondant a un polygone ou un point donne. Les coordonnees X et
Y en Lambert II étendu des mailles Meteo France ont ete fusionnee avec le
GeoPackage des mailles Safran de Bertuzzi et Clastre, 2022.
}
\author{
Baptiste ANTOINE
}
