% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim2.R
\name{get_sim2}
\alias{get_sim2}
\title{Extract SIM2 climate data}
\source{
Meteo France -> meteo.data.gouv.fr
}
\usage{
get_sim2(years, x, y)
}
\arguments{
\item{years}{-> Sélection des années ; maximum 30}

\item{x}{-> Coordonnees X de la maille en LAMBERT}

\item{y}{-> Coordonnees Y de la maille en LAMBERT}
}
\value{
La fonction renvoie un data frame contenant les 28 vaiables du modèle Safran-Isba-Modcou de Meteo France
}
\description{
Cette fonction permet d'extraire pour une maille donnee les
donnees fournies par Meteo France dans son modele de surface Safran-Isba-Modcou.
La periode d'extraction est limitee a 30 ans pour des questions de RAM.
}
\author{
Baptiste ANTOINE
}
