% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maille_sim2.R
\docType{data}
\name{maille_sim2}
\alias{maille_sim2}
\title{Information sur les mailles Safran.}
\format{
data frame contenant 8602 lignes et 25 variables.
}
\source{
Meteo France (<https://meteo.data.gouv.fr/>) and Bertuzzi, Patrick; Clastre, Philippe, 2022, "Information sur les mailles SAFRAN", DOI 10.57745/1PDFNL, Recherche Data Gouv, V2.
}
\usage{
maille_sim2
}
\description{
Table permettant de visualiser et extraire les mailles Safran pour une zone geographique donnee.
}
\examples{
data(maille_sim2)

}
\keyword{datasets}
